requirements.txt
requests>=2.31.0
python-dateutil>=2.8.2
tabulate>=0.9.0
aiogram>=3.3.0
